Refactor: Extract source image embedding and video metadata logic into separate modules
- Created shared/utils/source_image_embedding.py (~650 lines)
  - All MP4/MKV source image embedding logic (mutagen + ffmpeg)
  - All source image extraction logic
  - Helper functions: prepare_source_images_dict, process_extracted_images, apply_images_to_settings
  - Complete orchestration function: extract_and_apply_to_settings

- Created shared/utils/video_metadata.py (~165 lines)
  - MP4/MKV metadata read/write functions
  - Uses mutagen for MP4, ffmpeg for MKV
  - Clean API: save_metadata_to_video, read_metadata_from_video

- Refactored wgp.py (~815 lines removed)
  - Removed all MP4/MKV handling business logic
  - Reduced to minimal UI glue (get state, call module, set state)
  - extract_and_apply_source_images now just 6 lines

- Updated audio_video.py to import from new modules

Result: Clean separation of UI layer vs business logic, improved testability and reusability
